[#15] feat : 건강수첩, 에디터 페이지 기능 - 세부 수정
