+++ components/@common/HealthLogTaskList/style.css.ts
@@ -1,0 +1,10 @@
+import { style } from "@vanilla-extract/css";
+
+export const date = style({
+  padding: "1rem 0",
+
+  backgroundColor: "lightgoldenrodyellow",
+
+  fontSize: "1.4rem",
+  color: "gray",
+});
