+++ package.json
@@ -32,6 +32,7 @@
     "axios": "^1.6.5",
     "eslint": "^8",
     "eslint-config-next": "14.0.4",
+    "eslint-import-resolver-typescript": "^3.6.1",
     "husky": "^8.0.3",
     "lint-staged": "^15.2.0",
     "prettier": "^3.2.4",
